+++ project/dbt/{{ cookiecutter.project_name }}/resources/{{ cookiecutter.project_name }}.tf
@@ -1,18 +1,3 @@
-locals {
-  service_account_name = "{{ cookiecutter.project_name.replace('_', '-') }}"
-}
-
-resource "kubernetes_service_account" "default" {
-  metadata {
-    name = local.service_account_name
-    namespace = var.env_name
-    annotations = {
-      "eks.amazonaws.com/role-arn" : {{ '"{}"'.format(cookiecutter.role_arn) if cookiecutter.role_creation == 'external' else 'aws_iam_role.default.arn' }}
-    }
-  }
-}
-
-{%- if cookiecutter.role_creation == "project" %}
 resource "aws_iam_role" "default" {
   name               = "{{ cookiecutter.project_name }}-${var.env_name}"
   assume_role_policy = data.aws_iam_policy_document.default.json
@@ -26,7 +11,7 @@ data "aws_iam_policy_document" "default" {
     condition {
       test     = "StringEquals"
       variable = "${replace(var.aws_iam_openid_connect_provider_url, "https://", "")}:sub"
-      values   = ["system:serviceaccount:${var.env_name}:${local.service_account_name}"]
+      values   = ["system:serviceaccount:${var.env_name}:{{ cookiecutter.project_name }}-*"]
     }
 
     principals {
@@ -35,4 +20,3 @@ data "aws_iam_policy_document" "default" {
     }
   }
 }
-{%- endif %}