+++ tests/projects/test_python.py
@@ -36,7 +36,7 @@ def test_python_template_pipenv(cookies):
 def test_python_template_no_role(cookies):
     result = cookies.bake(
         template=f"{os.path.dirname(os.path.abspath(__file__))}/../../project/python",
-        extra_context={"role_creation": "none"},
+        extra_context={"datafy_managed_role": False},
     )
     assert 0 == result.exit_code
     assert result.exception is None